% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherExactTest.R
\name{fisherExactTest}
\alias{fisherExactTest}
\title{Fisher Exact Test}
\usage{
fisherExactTest(DataTab, alpha = 0.05)
}
\arguments{
\item{DataTab}{A 1-way frequency table as produced by \code{xtabs()}.}

\item{alpha}{A real number between 0 and 1.}
}
\value{
\code{fisherExactTest()} will return a named list of following values:
   \itemize{
   \item \code{pval} p value indicating the reslult of hypothesis test.
   \item \code{data} The passed in data with added margin sums.
   }
   A bar plot will also be generated that plots the row proportions of the data
   so we can better understand the data.
}
\description{
The function \code{fisherExactTest()} takes a 2-way, 2xC frequency table and
   some value \eqn{\alpha} to conduct a hypothesis test of independence on the data
   A plot (inspired by `s20x::rowdistr`) is made to better understand the data. The
   Main function used is `fisher.test()` which provides the p-value of our test.
}
