% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneWay.R
\name{oneWay}
\alias{oneWay}
\title{Oneway Table Analysis}
\usage{
oneWay(DataTab, alpha = 0.05)
}
\arguments{
\item{DataTab}{A 1-way frequency table as produced by \code{xtabs()}.}

\item{alpha}{A real number between 0 and 1.}
}
\value{
\code{oneWay()} will return a named list of following values:
   \itemize{
   \item \code{prob} the estimation for \eqn{p}.
   \item \code{Conf} confidence interval data for each \eqn{p}.
   \item \code{Diff} differences between each \eqn{p}.
   }
   A bar plot will also be generated that plots the values of \eqn{p} along with
   their respective confidence interval along with a horizontal line at
   \code{h = 1/length{DataTab}}. Each bar is shaded according to its \eqn{p} value;
   the larger \eqn{p} value corresponds to a darker bar.

   A scatter plot and intervals are also returned.
}
\description{
The function \code{oneWay()} takes a 1-way frequency table and
   some value \eqn{\alpha} to create a \eqn{(1-\alpha)100\%} confidence
   interval on each estimate \eqn{p} from the table. A plot (inspired by
   `s20x`) is made to show each \eqn{p} value along with their respective
   confidence intervals and the null \eqn{p_0} value. A similar process will
   be run on the differences of each proportion
}
\details{
The values for \eqn{p_i} are calculated as
   \eqn{\frac{\# \text{ in group } i}{n}}. Each confidence interval is
   made from a lower and upper bound calculated from Wald's method:

   \eqn{p \pm z_{\alpha/2}\sqrt{\frac{p(1-p)}{n}}}
}
\examples{
## This is a contingency table in array form
DF <- as.data.frame(UCBAdmissions)

## Make Cross Table
tab <- xtabs(Freq ~ Dept, DF)

## Run the Function
oneWay(tab, alpha = 0.05)
}
