% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoWay.R
\name{twoWay}
\alias{twoWay}
\title{Two Way Table Analysis}
\usage{
twoWay(DataTable, alpha = 0.05)
}
\arguments{
\item{DataTable}{A 1-way frequency table as produced by \code{xtabs()}.}

\item{alpha}{A real number between 0 and 1.}
}
\value{
\code{twoWay()} will return a named list of following values:
   \itemize{
   \item \code{Props} the estimation for \eqn{p} on each row.
   \item \code{Counts} counts of each occurance.
   \item \code{Expected} expected count of each data, generated by \code{chisq.test}.
   \item \code{RowProps} \eqn{p} by row
   \item \code{pval} p-value to accept or reject null.
   }
   A bar plot will also be generated that plots the row proportions along with
   their respective confidence interval. Each bar is shaded according to
   its row.
}
\description{
The function \code{twoWay()} takes a 2-way frequency table and
   some value \eqn{\alpha} to study. A plot (inspired by
   `s20x::rowdistr`) is created to help compare the data and better
   understand if there is independece between the data.
}
